# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
      "AggregatorEntry": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "keywords": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "source": {
            "enum": [
              "remoteok",
              "hn-hiring",
              "themuse",
              "remotive",
              "adzuna",
              "jooble",
              "jsearch"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "AggregatorEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResumeBase": {
        "additionalProperties": false,
        "properties": {
          "block_ids": {
            "items": {
              "type": "string"
            },
            "title": "Block Ids",
            "type": "array"
          },
          "id": {
            "pattern": "^[a-z0-9][a-z0-9-]*$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "section_order": {
            "default": [
              "summary",
              "experience",
              "projects",
              "skills",
              "credentials"
            ],
            "items": {
              "type": "string"
            },
            "title": "Section Order",
            "type": "array"
          },
          "style": {
            "additionalProperties": true,
            "default": {},
            "title": "Style",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "block_ids"
        ],
        "title": "ResumeBase",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Rhapto API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/profile/aggregators": {
      "get": {
        "operationId": "get_aggregators_api_v1_profile_aggregators_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AggregatorEntry"
                  },
                  "title": "Response Get Aggregators Api V1 Profile Aggregators Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Aggregators",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "operationId": "put_aggregators_api_v1_profile_aggregators_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/AggregatorEntry"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AggregatorEntry"
                  },
                  "title": "Response Put Aggregators Api V1 Profile Aggregators Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Aggregators",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile/answers": {
      "get": {
        "operationId": "get_answers_api_v1_profile_answers_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Answers Api V1 Profile Answers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Answers",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "operationId": "put_answers_api_v1_profile_answers_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Put Answers Api V1 Profile Answers Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Answers",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile/bases": {
      "get": {
        "operationId": "list_bases_api_v1_profile_bases_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumeBase"
                  },
                  "title": "Response List Bases Api V1 Profile Bases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bases",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile/bases/{base_id}": {
      "delete": {
        "operationId": "delete_base_api_v1_profile_bases__base_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Base",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "operationId": "put_base_api_v1_profile_bases__base_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeBase"
                }
              }
            },
            "description": "The base already existed and was replaced"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeBase"
                }
              }
            },
            "description": "The base was created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Base",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile/blocks": {
      "get": {
        "operationId": "list_blocks_api_v1_profile_blocks_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Block"
                  },
                  "title": "Response List Blocks Api V1 Profile Blocks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Blocks",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile/blocks/{block_id}": {
      "delete": {
        "operationId": "delete_block_api_v1_profile_blocks__block_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "block_id",
            "required": true,
            "schema": {
              "title": "Block Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Block",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "operationId": "put_block_api_v1_profile_blocks__block_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "block_id",
            "required": true,
            "schema": {
              "title": "Block Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Block"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            },
            "description": "The block already existed and was replaced"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            },
            "description": "The block was created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Block",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/profile/export": {
      "get": {
        "operationId": "export_profile_api_v1_profile_export_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Profile",
        "tags": [
          "profile"
        ]
      }
    },
...
  }
}